backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
python-multipart==0.0.12
websockets==13.1
opencv-python==4.10.0.84
easyocr==1.7.1
sqlalchemy==2.0.36
aiohttp==3.10.10
httpx==0.27.2
google-generativeai==0.8.3
pillow==11.0.0
numpy==1.26.4
